Create wifigrabberv2 for ducky
this is designed for windows 10 to grab all known wifi connections, dump them to the t folder then move them to said "t" folder on the flash drive. you have to change 'ducky' to the name of your drive. you will need to leave the ' symbols. if you want to see it run remove the -windowstyle hidden toward the top.